Move nock logic to TestDestination class
