+++ packages/cli/templates/actions/snapshot/__tests__/snapshot.test.ts
@@ -1,7 +1,4 @@
-import { createTestEvent, createTestIntegration } from '@segment/actions-core'
-import { Settings } from '../generated-types'
-import nock from 'nock'
-import generateTestData from '@segment/action-destinations/src/lib/test-data'
+import { createTestIntegration } from '@segment/actions-core'
 import destination from '../index'
 
 const testDestination = createTestIntegration(destination)
@@ -9,24 +6,7 @@ const testDestination = createTestIntegration(destination)
 describe('Testing snapshot for {{destination}} action:', () => {
   for (const actionSlug in destination.actions) {
     it(`${actionSlug} [required]`, async () => {
-      const action = destination.actions[actionSlug]
-      const [eventData, settingsData] = generateTestData(destination, action, 'required')
-
-      nock(/.*/).persist().get(/.*/).reply(200)
-      nock(/.*/).persist().post(/.*/).reply(200)
-      nock(/.*/).persist().put(/.*/).reply(200)
-
-      const event = createTestEvent({
-        properties: eventData
-      })
-
-      const responses = await testDestination.testAction(actionSlug, {
-        event,
-        settings: settingsData as Settings,
-        mapping: event.properties,
-        useDefaultMappings: false
-      })
-
+      const responses = await testDestination.snapshotAction(destination, actionSlug, true)
       const requestBody = responses[0].request.body
       if (requestBody) {
         expect(requestBody.toString()).toMatchSnapshot()
@@ -34,24 +14,7 @@ describe('Testing snapshot for {{destination}} action:', () => {
     })
 
     it(`${actionSlug} [all fields]`, async () => {
-      const action = destination.actions[actionSlug]
-      const [eventData, settingsData] = generateTestData(destination, action, 'all')
-
-      nock(/.*/).persist().get(/.*/).reply(200)
-      nock(/.*/).persist().post(/.*/).reply(200)
-      nock(/.*/).persist().put(/.*/).reply(200)
-
-      const event = createTestEvent({
-        properties: eventData
-      })
-
-      const responses = await testDestination.testAction(actionSlug, {
-        event,
-        settings: settingsData as Settings,
-        mapping: event.properties,
-        useDefaultMappings: false
-      })
-
+      const responses = await testDestination.snapshotAction(destination, actionSlug, false)
       const requestBody = responses[0].request.body
       if (requestBody) {
         expect(requestBody.toString()).toMatchSnapshot()
